requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
imageio==2.9.0
imgpy==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
python-dotenv==0.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
